+++ src/Domain/Collection/Events.php
@@ -3,7 +3,7 @@
 /*
  * This file is part of the eluceo/iCal package.
  *
- * (c) 2022 Markus Poerschke <markus@poerschke.nrw>
+ * (c) 2023 Markus Poerschke <markus@poerschke.nrw>
  *
  * This source file is subject to the MIT license that is bundled
  * with this source code in the file LICENSE.
@@ -16,6 +16,9 @@ use Iterator;
 use IteratorAggregate;
 use ReturnTypeWillChange;
 
+/**
+ * @implements IteratorAggregate<Event>
+ */
 abstract class Events implements IteratorAggregate
 {
     /**