+++ src/Domain/Entity/Calendar.php
@@ -3,7 +3,7 @@
 /*
  * This file is part of the eluceo/iCal package.
  *
- * (c) 2022 Markus Poerschke <markus@poerschke.nrw>
+ * (c) 2023 Markus Poerschke <markus@poerschke.nrw>
  *
  * This source file is subject to the MIT license that is bundled
  * with this source code in the file LICENSE.
@@ -11,6 +11,7 @@
 
 namespace Eluceo\iCal\Domain\Entity;
 
+use DateInterval;
 use Eluceo\iCal\Domain\Collection\Events;
 use Eluceo\iCal\Domain\Collection\EventsArray;
 use Eluceo\iCal\Domain\Collection\EventsGenerator;
@@ -20,6 +21,8 @@ use Iterator;
 class Calendar
 {
     private string $productIdentifier = '-//eluceo/ical//2.0/EN';
+
+    private ?DateInterval $publishedTTL = null;
 
     private Events $events;
 
@@ -54,6 +57,18 @@ class Calendar
         }
 
         throw new InvalidArgumentException('$events must be an array, an object implementing Iterator or an instance of Events.');
+    }
+
+    public function getPublishedTTL(): ?DateInterval
+    {
+        return $this->publishedTTL;
+    }
+
+    public function setPublishedTTL(?DateInterval $ttl): self
+    {
+        $this->publishedTTL = $ttl;
+
+        return $this;
     }
 
     public function getProductIdentifier(): string