+++ src/Domain/ValueObject/EmailAddress.php
@@ -3,7 +3,7 @@
 /*
  * This file is part of the eluceo/iCal package.
  *
- * (c) 2022 Markus Poerschke <markus@poerschke.nrw>
+ * (c) 2023 Markus Poerschke <markus@poerschke.nrw>
  *
  * This source file is subject to the MIT license that is bundled
  * with this source code in the file LICENSE.
@@ -33,6 +33,6 @@ final class EmailAddress
 
     public function toUri(): Uri
     {
-        return new Uri('mailto:' . urlencode($this->emailAddress));
+        return new Uri('mailto:' . $this->emailAddress);
     }
 }