+++ src/Presentation/Factory/CalendarFactory.php
@@ -3,7 +3,7 @@
 /*
  * This file is part of the eluceo/iCal package.
  *
- * (c) 2022 Markus Poerschke <markus@poerschke.nrw>
+ * (c) 2023 Markus Poerschke <markus@poerschke.nrw>
  *
  * This source file is subject to the MIT license that is bundled
  * with this source code in the file LICENSE.
@@ -14,6 +14,7 @@ namespace Eluceo\iCal\Presentation\Factory;
 use Eluceo\iCal\Domain\Entity\Calendar;
 use Eluceo\iCal\Presentation\Component;
 use Eluceo\iCal\Presentation\Component\Property;
+use Eluceo\iCal\Presentation\Component\Property\Value\DurationValue;
 use Eluceo\iCal\Presentation\Component\Property\Value\TextValue;
 use Generator;
 
@@ -56,5 +57,10 @@ class CalendarFactory
         yield new Property('VERSION', new TextValue('2.0'));
         /* @see https://www.ietf.org/rfc/rfc5545.html#section-3.7.1 */
         yield new Property('CALSCALE', new TextValue('GREGORIAN'));
+        $publishedTTL = $calendar->getPublishedTTL();
+        if ($publishedTTL) {
+            /* @see http://msdn.microsoft.com/en-us/library/ee178699(v=exchg.80).aspx */
+            yield new Property('X-PUBLISHED-TTL', new DurationValue($publishedTTL));
+        }
     }
 }