+++ CHANGELOG.md
@@ -7,9 +7,45 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added
+### Changed
 
 -   Declare CalendarFactory::getProperties protected i/o private to open it for extension
+
+## [2.9.0] - 2022-12-23
+
+### Changed
+
+-   Add missing return types and missing template annotations [#472](https://github.com/markuspoerschke/iCal/pull/472)
+    -   Added `@implements IteratorAggregate<Event>` to `\Eluceo\iCal\Domain\Collection\Events`
+    -   Added `@implements IteratorAggregate<ContentLine>` to `\Eluceo\iCal\Presentation\Component`
+    -   Added `@return Traversable<ContentLine>` to `\Eluceo\iCal\Presentation\Component::getIterator`
+-   Fix EmailAddress value object: do not url encode email addresses [#479](https://github.com/markuspoerschke/iCal/pull/479)
+
+## [2.8.0] - 2022-12-22
+
+### Added
+
+-   Support chaining of `Calendar::setPublishedTTL()` [#452](https://github.com/markuspoerschke/iCal/pull/452)
+-   Support PHP 8.2 [#470](https://github.com/markuspoerschke/iCal/pull/470)
+
+## [2.7.0] - 2022-06-21
+
+### Added
+
+-   Support `X-PUBLISHED-TTL` property on calendars [#413](https://github.com/markuspoerschke/iCal/pull/413)
+
+## [2.6.0] - 2022-06-17
+
+### Added
+
+-   Support status property on events [#422](https://github.com/markuspoerschke/iCal/pull/422)
+-   Support categories property on events [#421](https://github.com/markuspoerschke/iCal/pull/421)
+
+## [2.5.1] - 2022-04-26
+
+### Fixed
+
+-   Deprecation message occuring on an OOTB Symfony 5.4 on PHP 7.4 [#382](https://github.com/markuspoerschke/iCal/pull/382)
 
 ## [2.5.0] - 2022-02-13
 
@@ -220,7 +256,12 @@ Please check the [upgrade guide](UPGRADE.md) on how to upgrade from version `0.*
 -   **Breaking Change:** Changed signature of the `Event::setOrganizer` method. Now there is is only one parameter that must be an instance of `Property\Organizer`.
 -   Updated install section in README.md [#54](https://github.com/markuspoerschke/iCal/pull/53)
 
-[unreleased]: https://github.com/markuspoerschke/iCal/compare/2.5.0...HEAD
+[unreleased]: https://github.com/markuspoerschke/iCal/compare/2.9.0...HEAD
+[2.9.0]: https://github.com/markuspoerschke/iCal/compare/2.8.0...2.9.0
+[2.8.0]: https://github.com/markuspoerschke/iCal/compare/2.7.0...2.8.0
+[2.7.0]: https://github.com/markuspoerschke/iCal/compare/2.6.0...2.7.0
+[2.6.0]: https://github.com/markuspoerschke/iCal/compare/2.5.1...2.6.0
+[2.5.1]: https://github.com/markuspoerschke/iCal/compare/2.5.0...2.5.1
 [2.5.0]: https://github.com/markuspoerschke/iCal/compare/2.4.0...2.5.0
 [2.4.0]: https://github.com/markuspoerschke/iCal/compare/2.3.0...2.4.0
 [2.3.0]: https://github.com/markuspoerschke/iCal/compare/2.2.0...2.3.0